+++ app/login/page.tsx
@@ -3,7 +3,7 @@ import AuthButtons from '@/components/Auth/AuthButtons'
 import AuthInput from '@/components/Auth/AuthInput'
 import AuthTitle from '@/components/Auth/AuthTitle'
 import Modal from '@/components/common/Modal'
-import useAuthForm from '@/hooks/useAuthForm'
+import useAuthForm from '@/hooks/useForm'
 import useModal from '@/hooks/useModal'
 import { login } from '@/services/auth'
 import { useAuthStore } from '@/store/useAuthStore'
